Preview rendered version of an Rd file

Function to preview the rendered version of an Rd file.

preview_rd(Rdfile, view = TRUE, type = "html", verbose = FALSE)

Arguments

Rdfile
character string with the name of the Rd file to preview (either with or without the .Rd or .rd extension).

view
logical indicating whether the rendered version of the help file should be displayed.

type
character string indicating which version should be rendered (either "html", "txt", or "pdf").

verbose
logical indicating whether diagnostic output will be shown.

Details

The function is useful when writing a help file that contains MathJax equations. Instead of having to reinstall the package under development to check if the equations are being rendered correctly, one can just set the current working directory to the root of the package (or its man directory) and then use preview_rd() to preview the HTML, plain-text, or PDF version of an Rd file on the fly.

For type="html", the HTML page will be opened in the browser with the browseURL function. When making further changes to the Rd file, reopening the page each time preview_rd() is called is inconvenient as this will usually open up a new tab in the browser. Setting view=FALSE prevents this. Reloading the page in the open tab should then reflect the updates. In RStudio, the generated HTML version will be displayed in the ‘Viewer’ pane and the view argument is then irrelevant.

For type="txt", the plain-text version of the help file will be shown (using the file.show function and directly on the console in RStudio).

For type="pdf", the PDF is generated using R CMD Rd2pdf and should open up in the default PDF viewer.

Due to some limitations as to how MathJax can be loaded via the mathjaxr package, MathJax must be loaded via the CDN. Hence, rendering of equations in HTML will only work with an active internet connection.